backend/requirements.txt
daphne>=4.0.0
channels>=4.0.0
channels_redis>=4.1.0
Django>=5.0,<6.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
psycopg2-binary>=2.9.9
django-cors-headers>=4.3.0
python-dotenv>=1.0.0
redis>=5.0.0
django-redis>=5.4.0
drf-spectacular>=0.27.0
Pillow>=10.0.0
gunicorn>=21.2.0
celery>=5.3.0
django-storages>=1.14
boto3>=1.34
django-filter>=23.5

